Allow creation of multiple keystone instances
Inject keystone instance into  admin/server/api/cloudinary

Inject lists into admin/server/api/counts
