+++ admin/server/app/createDynamicRouter.js
@@ -14,8 +14,8 @@ module.exports = function createDynamicRouter(keystone) {
 
 	// Use bodyParser and multer to parse request bodies and file uploads
 	router.use(bodyParser.json({}));
-	router.use(bodyParser.urlencoded({ extended: true }));
-	router.use(multer({ includeEmptyFields: true }));
+	router.use(bodyParser.urlencoded({extended: true}));
+	router.use(multer({includeEmptyFields: true}));
 
 	// #1: Session API
 	// TODO: this should respect keystone auth options
@@ -50,9 +50,10 @@ module.exports = function createDynamicRouter(keystone) {
 	// #4: Cloudinary and S3 specific APIs
 	// TODO: poor separation of concerns; should / could this happen elsewhere?
 	if (keystone.get('cloudinary config')) {
-		router.get('/api/cloudinary/get', require('../api/cloudinary').get);
-		router.get('/api/cloudinary/autocomplete', require('../api/cloudinary').autocomplete);
-		router.post('/api/cloudinary/upload', require('../api/cloudinary').upload);
+		var cloudinary = require('../api/cloudinary')(keystone);
+		router.get('/api/cloudinary/get', cloudinary.get);
+		router.get('/api/cloudinary/autocomplete', cloudinary.autocomplete);
+		router.post('/api/cloudinary/upload', cloudinary.upload);
 	}
 	if (keystone.get('s3 config')) {
 		router.post('/api/s3/upload', require('../api/s3').upload);
@@ -67,7 +68,7 @@ module.exports = function createDynamicRouter(keystone) {
 	// Init req with list
 	var initList = require('../middleware/initList')(keystone);
 
-	router.all('/api/counts', require('../api/counts'));
+	router.all('/api/counts', require('../api/counts')(keystone.lists));
 	router.get('/api/:list', initList(), require('../api/list/get'));
 	router.get('/api/:list/:format(export.csv|export.json)', initList(), require('../api/list/download'));
 	router.post('/api/:list/delete', initList(), require('../api/list/delete'));
