+++ lib/list.js
@@ -12,7 +12,6 @@ var utils = require('keystone-utils');
 
 function List(key, options) {
 	if (!(this instanceof List)) return new List(key, options);
-
 	var defaultOptions = {
 		schema: {
 			collection: keystone.prefixModel(key)
